@@ -3,21 +3,21 @@
   "version": "1.0.0",
   "description": "Javascript LDPC codec",
   "main": "index.js",
+  "type": "module",
   "directories": {
     "test": "test"
   },
-  "dependencies": {},
   "devDependencies": {
-    "chai": "^4.2.0",
-    "eslint": "^7.8.0",
-    "mathjs": "^7.2.0",
-    "mocha": "^8.1.3",
-    "rollup": "^2.26.8"
+    "chai": "^4.3.8",
+    "eslint": "^8.49.0",
+    "mathjs": "^11.11.0",
+    "mocha": "^10.2.0",
+    "rollup": "^3.29.2"
   },
   "scripts": {
     "test": "node node_modules/mocha/bin/mocha",
     "lint": "node node_modules/eslint/bin/eslint src/*.js test/*.js",
-    "build": "node node_modules/rollup/dist/bin/rollup -c"
+    "build": "node node_modules/rollup/dist/bin/rollup -c -f es"
   },
   "repository": {
     "type": "git",
