@@ -2,6 +2,6 @@ export default {
 	input: 'src/index.js',
 	output: {
 	  file: 'dist/ldpc802.js',
-	  format: 'cjs'
+	  format: 'module'
 	}
   };
